add migrations table and create migrate.py script
