+++ api/Dockerfile
@@ -6,6 +6,6 @@ COPY ./requirements.txt /api/requirements.txt
 
 RUN pip install --no-cache-dir --upgrade -r /api/requirements.txt
 
-COPY ./src /api/src
+COPY . .
 
 CMD ["fastapi", "run", "src/main.py", "--port", "3000"]
